requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
mysql-connector-python==8.3.0
requests==2.31.0
pydantic==2.6.4
bcrypt==4.1.2